+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## Currently Unreleased - TBD
 
 * Support registering member functions with noexcept-specification for C++17 or later.
+* Support binding C++ functions with noexcept-specification for C++17 or later.
 
 ## v1.2.1 - 2024.10.15